cleaned up storyboard loading in app delegate

+++ GadaboutiOS/LocationController.m
@@ -83,7 +83,7 @@
 - (void)getAuthorization {
     if(IS_OS_8_OR_LATER) {
         NSUInteger authCode = [CLLocationManager authorizationStatus];
-        NSLog(@"Requesting location tracking permission; Auth status: %d", authCode);
+        NSLog(@"Requesting location tracking permission; Auth status: %lu", (unsigned long)authCode);
         if ((unsigned long)authCode == kCLAuthorizationStatusNotDetermined && [locationManager respondsToSelector:@selector(requestWhenInUseAuthorization)]) {
             [locationManager requestWhenInUseAuthorization];
         }
